BDOG-1119 Move HeaderCarrierConverter from uk.gov.hmrc.play into uk.gov.hmrc.play.http
