+++ http-verbs-play-26/src/main/scala/uk/gov/hmrc/package.scala
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2020 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc
+
+package play {
+
+  @deprecated("Use uk.gov.hmrc.play.http.HeaderCarrierConverter instead", "13.0.0")
+  object HeaderCarrierConverter extends uk.gov.hmrc.play.http.HeaderCarrierConverter
+}
